Now works properly in docker using env variables

+++ startup.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
-if [ -f /config/ClassifierParser.class]
+if [ -f "/config/ClassifierParser.class" ] && [ cmp "/config/ClassifierParser.class" "/app/src/main/java/Classifier/classifierParser.java" ];
 then
 	mv -f /config/ClassifierParser.java /app/src/main/java/Classifier/ClassifierParser.java
+	cd /app
+	mvn clean compile assembly:single
+	cd /
+elif [ ! -f "/app/target/weka-service-1.0-SNAPSHOT-jar-with-dependencies.jar" ];
+then
+	cd /app
+	mvn clean compile assembly:single
+	cd /
 fi
-cd /app
-mvn clean compile assembly:single
 cd /app/target
 java -cp weka-service-1.0-SNAPSHOT-jar-with-dependencies.jar Main.Main
